+++ doc/whatsnew.rst
@@ -32,6 +32,8 @@ v0.10.dev
 
 - Add :func:`pyriemann.utils.geodesic.geodesic_chol` and :func:`pyriemann.utils.mean.mean_chol`. :pr:`399` by :user:`qbarthelemy`
 
+- Add :func:`pyriemann.utils.tangentspace.transport_logchol` for parallel transport with log-Cholesky metric. :pr:`400` by :user:`qbarthelemy`
+
 v0.9 (July 2025)
 ----------------